57c

+++ ABC/57/c.cpp
@@ -3,8 +3,8 @@
 
 using namespace std;
 
-int calc(long int a, long int b){
-    long int la = log10(a), lb = log10(b);
+int calc(int a, int b){
+    int la = log10(a), lb = log10(b);
     if(la >= lb) return la+1;
     else return lb+1;
 }
@@ -16,7 +16,7 @@ int main(void){
     int rn = sqrt(n);
     for(double i=1; i<=rn; ++i){
         double j = n / i;
-        if(j == (long int)j && r > calc(i, j)) r = calc(i, j);
+        if(j == (int)j && r > calc(i, j)) r = calc(i, j);
     }
     cout << r << endl;
     return 0;
